Move unused testing functions to proper testing file
